feat: add cli flags for storage capacity per subnetwork

+++ trin-state/src/storage.rs
@@ -222,7 +222,7 @@ pub mod test {
 
     use super::*;
 
-    const STORAGE_CAPACITY_MB: u64 = 10;
+    const STORAGE_CAPACITY_MB: u32 = 10;
 
     struct ContentData {
         key: StateContentKey,
